Add files via upload
every refresh draws 101 circles. 100 blue 1 white. there are some bars too. i used them for coordinate reference. the goal is to learn to make a game in a sectioned canvas. i could make a counter and event listener and action, click or tap/swipe-across white circle for a point, have it auto refresh over and over for the duration of the game.